Made StackObjectPool's maxIdle parameter configurable in RedisClientPool
The maxIdle in StackObjectPool determines that when an object is returned
to the pool and there is more than maxIdle objects idle, then the bottom
object is destroyed and the returned object is inserted at the top of
the stack. With its default value of 8 and enough concurrent usage (more
than 8 threads) you can easily start creating and destroying a lot of
objects in sequence.

In high concurrency contexts one can increase maxIdle to the number of
threads to avoid this unnecessary destroying/creating.

The high number of creating/destroying can lead to exhausting ports
causing:
java.lang.RuntimeException: java.net.NoRouteToHostException: Cannot
assign requested address